bitwise class tests

+++ src/sim/gate/playground/TestParser.java
@@ -14,7 +14,7 @@ public class TestParser implements ITest {
     }
 
     private void scenario1() {
-        byte[] inputString = {1, 10, 71, 0, 10, 7, 0, 1};
+        int[] inputString = {1, 10, 71, 0, 10, 7, 0, 1};
 
         CCS2Configuration conf = new CCS2Configuration();
         conf.inputs = 2;
